feat: Add MAX_SPEED constant for motor speed control
The code changes add a new constant `MAX_SPEED` with a value of 60 to control the maximum speed of the motor. This constant will be used in the `PiStepper` class to limit the speed set by the `setSpeed` function.

Note: This suggested commit message follows the established convention of using a prefix to indicate the type of change (feat for feature).